requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.26.0
faiss-cpu>=1.7.4
networkx>=3.1
sentence-transformers>=2.2.2
readerwriterlock>=1.0.9
scikit-learn>=1.3.0
streamlit>=1.28.0
requests>=2.31.0
matplotlib>=3.8.0
python-multipart>=0.0.6

# Ingestion Pipeline Dependencies
markdown-it-py>=3.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
nltk>=3.8.0
